Adds TweakPanels to Pottery app
An app can add a bunch of "tweak-panel" class divs with a bunch of tweaks.  Then you can create a select to pick one of them at a time.  The tweaks are all still present even when they're hidden, so it can be used to compartmentalize the list of tweaks.

Adds a patternType select so that each type can have its own list of tweaks, and switching the type will recreate the geometry accordingly

Adds a new clover pattern type that currently doesn't have any tweaks and simply draws a single hole at 0, 0